Sun Jan 20 11:07:38 EST 2008
----------------------------
svn merge -r 604350:613587 https://svn.apache.org/repos/asf/directory/apacheds/b
ranches/bigbang .



git-svn-id: https://svn.apache.org/repos/asf/directory/sandbox/akarasulu/bigbang/apacheds@613608 13f79535-47bb-0310-9956-ffa450edef68

+++ protocol-dns/src/main/java/org/apache/directory/server/dns/messages/RecordClass.java
@@ -56,7 +56,7 @@ public enum RecordClass implements EnumConverter<Short>
 
     private RecordClass( int value )
     {
-        this.value = ( byte ) value;
+        this.value = ( short ) value;
     }
